+++ docs/conf.py
@@ -9,7 +9,7 @@
 project = 'pydantic-resolve'
 copyright = '2023, tangkikodo'
 author = 'tangkikodo'
-release = '0.5.0'
+release = '1.1.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration